Rework of setup logic for easier scaling in future.

+++ scripts/companion-setup.sh
@@ -1,16 +1,24 @@
 #!/usr/bin/env bash
 
-#Get install script install.sh from github.
-curl -O https://raw.githubusercontent.com/bitfocus/companion-pi/main/install.sh
+# Asks to run script
+echo -e -n  "\n\033[1m"Install Companion?"\033[0m (y/n): "
+    read -r -p "" choice < /dev/tty
+    if [[ "$choice" =~ ^[Yy]$ ]]; then
 
-#Make the script executable
-chmod +x install.sh
+	# Get install script install.sh from github.
+	curl -O https://raw.githubusercontent.com/bitfocus/companion-pi/main/install.sh
 
-#Sets stable build and runs the script with root privileges using bash
-sudo COMPANION_BUILD="v3.5.4" bash install.sh
+	# Make the script executable
+	chmod +x install.sh
 
-#Removes the script from show-pi directory
-rm install.sh
+	# Sets stable build and runs the script with root privileges using bash
+	sudo COMPANION_BUILD="v3.5.5" bash install.sh
 
-# Starts the compannion service
-sudo systemctl start companion.service
+	# Removes the script from show-pi directory
+	rm install.sh
+
+	# Starts the compannion service
+	sudo systemctl start companion.service
+    else
+	echo -e "Skipped Companion install\n"
+    fi
